# internal/docs/spec/swagger.yaml
basePath: /api/v1
definitions:
...
    github_com_Angle-HR_server_internal_apidoc.ErrorBody:
        properties:
            code:
                example: VALIDATION_ERROR
                type: string
            details:
                additionalProperties: {}
                type: object
            message:
                example: invalid request body
                type: string
        type: object
    github_com_Angle-HR_server_internal_apidoc.ErrorEnvelope:
        properties:
            error:
                $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.ErrorBody'
            meta:
                $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.Meta'
        type: object
    github_com_Angle-HR_server_internal_apidoc.Meta:
        properties:
            has_more:
                type: boolean
            next_cursor:
                type: string
            request_id:
                example: abc123
                type: string
        type: object
...
    internal_handler.AdminWaitlistDetailEnvelope:
        properties:
            data:
                $ref: '#/definitions/internal_handler.adminWaitlistDetail'
            meta:
                $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.Meta'
        type: object
...
    internal_handler.AdminWaitlistPatchRequest:
        properties:
            metadata:
                items:
                    type: integer
                type: array
            wants_early_access:
                type: boolean
            wants_user_testing:
                type: boolean
        type: object
...
    internal_handler.SignupData:
        properties:
            message:
                example: You're on the list!
                type: string
            region:
                example: uk
                type: string
        type: object
    internal_handler.SignupEnvelope:
        properties:
            data:
                $ref: '#/definitions/internal_handler.SignupData'
            meta:
                $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.Meta'
        type: object
    internal_handler.SignupRequest:
        properties:
            email:
                example: jerry@example.com
                type: string
        type: object
...
    internal_handler.adminWaitlistDetail:
        properties:
            country_id:
                type: string
            created_at:
                type: string
            deleted_at:
                type: string
            email:
                type: string
            frustration_ids:
                items:
                    type: string
                type: array
            full_name:
                type: string
            hiring_tool_ids:
                items:
                    type: string
                type: array
            industry_ids:
                items:
                    type: string
                type: array
            metadata:
                items:
                    type: integer
                type: array
            onboarding_submitted_at:
                type: string
            region:
                type: string
            region_source:
                type: string
            role_id:
                type: string
            team_size_id:
                type: string
            updated_at:
                type: string
            uuid:
                type: string
            wants_early_access:
                type: boolean
            wants_user_testing:
                type: boolean
        type: object
...
host: localhost:8080
info:
    contact: {}
    description: |-
        Waitlist and product onboarding API for Open HR.
        Interactive docs (Scalar) are served at `/` in non-production environments.
        Product flow: signup → verify email (6-digit OTP, 5 min expiry) → profile → address → [business] → complete.
        Onboarding step endpoints require `Authorization: Bearer <access_token>` after email verification.
        Admin ops console APIs live under `/api/v1/admin` and require a separate admin JWT (`token_use=admin`).
    title: Angle HR API
    version: "1.0"
paths:
...
    /admin/waitlist/{uuid}:
        delete:
            parameters:
                - description: Waitlist UUID
                  in: path
                  name: uuid
                  required: true
                  type: string
                - description: Region hint
                  in: query
                  name: region
                  type: string
            responses:
                "204":
                    description: No Content
            security:
                - BearerAuth: []
            summary: Soft-delete waitlist entry
            tags:
                - admin/waitlist
        get:
            parameters:
                - description: Waitlist UUID
                  in: path
                  name: uuid
                  required: true
                  type: string
                - description: Region hint
                  in: query
                  name: region
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/internal_handler.AdminWaitlistDetailEnvelope'
                "404":
                    description: Not Found
                    schema:
                        $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.ErrorEnvelope'
            security:
                - BearerAuth: []
            summary: Get waitlist entry
            tags:
                - admin/waitlist
        patch:
            consumes:
                - application/json
            parameters:
                - description: Waitlist UUID
                  in: path
                  name: uuid
                  required: true
                  type: string
                - description: Region hint
                  in: query
                  name: region
                  type: string
                - description: Patch payload
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/internal_handler.AdminWaitlistPatchRequest'
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/internal_handler.AdminWaitlistDetailEnvelope'
            security:
                - BearerAuth: []
            summary: Update waitlist entry
            tags:
                - admin/waitlist
...
    /waitlist:
        post:
            consumes:
                - application/json
            description: Collects an email for the waitlist and writes regional plus global registry rows.
            parameters:
                - description: Signup payload
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/internal_handler.SignupRequest'
            produces:
                - application/json
            responses:
                "201":
                    description: Created
                    schema:
                        $ref: '#/definitions/internal_handler.SignupEnvelope'
                "400":
                    description: Bad Request
                    schema:
                        $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.ErrorEnvelope'
                "409":
                    description: Conflict
                    schema:
                        $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.ErrorEnvelope'
                "500":
                    description: Internal Server Error
                    schema:
                        $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.ErrorEnvelope'
            summary: Join waitlist
            tags:
                - waitlist/signup
securityDefinitions:
    BearerAuth:
        description: JWT access token. Send the raw token, or optionally prefix with "Bearer ".
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"
...